Fix issue 278
There are two issues with the quest Mom, The Adventurer that
this commit resolves. First, players with Bastok fame >=2 are
not offered this quest. Second, if the quest is accepted,
and a player's Bastok fame increases to 2 or higher, they
will be able to start the subsequent quest The Signpost Marks the Spot.
This MR fixes the logic of the NPC Nbu Latteh so the quest is offered
regardless of fame if it hasn't been completed, and requires that
Mom, The Adventurer is completed prior to offerring The Signpost Marks
The Spot. The existing behaviour of being able to repeat Mom, The Adventurer
so long as the player's Bastok fame is <2 is preserved.